+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   They can be created using `CFFIScope.create_function`.
 - Support for passing any packed array (other than PackedStringArray) as pointer to native functions.
 - `CFFIPointer.to_*_array` methods for copying data from pointers as other types of packed arrays
+- `StreamPeerCFFIPointer` stream peer that handles a binary data stream from a `CFFIPointer`.
 
 
 ## [0.1.0](https://github.com/gilzoide/cffi-gdextension/releases/tag/0.1.0)